Fix webpack config file extension and run script
